add KNNBrute variant for mathx.Distancer. + test and more.
Previously, the core KNN func was KNNBrute, which
operated on []float64 in a general-purpose way.
The usefulness of a variant which worked on
mathx.Distancer seemed plausable as well, so that
was implemented, along with prefab/wrapper funcs,
similar to KNNBrute.

Tests were implemented for this new feature (specifically
the prefab funcs, as those cover a lot), which are basically
re-formatted versions of tests written for KNNBrute prefabs/
wrappers.

Also, the TestKNNBrute test was deleted, it was pretty much
redundant, as the prefab tests covered the same areas.

Additionally, some documentation was fixed (spelling).