major refactoring and api change
Previously, there were two main/core KNN funcs: KNNBrute and
KNNBruteDist (for mathx.Distancer) with a signigicant amt of
code duplication.

This was solved by making it so that there is only one core,
a changed KNNBrute -- which now abstracts away the distance
calculation, and instead accepts a score generator in the arg.

Using this core, two conveniences are kept, one for []float64
gen and one for mathx.Distancer gen, so the functionality of
the pkg isn't reduced.

Also, this required a lot of updates in docs and tests.