changes to tests for previous commit
In the previous commit, some additional safety was added for when
the return of DistancerContainer.Distancer() was checked for nil.
This was because some external pkg implemented the DistancerContainer
interface in such a way that the ".Distancer()" method returned nil
after some expiration, which caused a panic in this pkg (knnc).
This scenario is now added in tests, which had some cascading changes.