optimize + change BaseWorkerArgs.BlockDeadline->TTL
Optimization:
Previously, all (concurrent) workers had a way of
cancelling implicitly through a deadline, which was
specified with BaseWorkerArgs.BlockDeadline. This
was a bit flawed/inconvenient in two ways: (1) the
duration would never be reduced. For instance, a
worker might run for a while and then get blocked;
at _that_ point, the duration would be waited. But
that isn't necessarily clear, as the duration is
arguably conceptualised more intutively as a TTL.
So a worker can't exist for more than that duration,
no matter what. As such, this is the first change.
The second (2) flaw was that the select cases in
all workers used the deadline poorly, they used
time.After(...) on each process iteration, which
is a costly upspin to be doing that frequently.
The solution here was to create a cancel signal
(that fires after the specified duration) outside
of the loops, and then reusing it. Doing a few
10s benchmarks showed that it was more than x10
faster (at least with SearchSpace.Scan).

Name change:
BaseWorkerArgs.BlockDeadline -> TTL. This required
some cascading changes in tests.

... also
Pipeline now uses BaseWorkerArgs too.