add latencytracker in pkg/timex
Some planned functionality will require the tracking of latency.
For instance, it will be useful to how much latency a repeated
process has had the last minute.

This tracking mechanism is based on a linked list (because it
trackes latency _the_last_period_), combined with cumulative
latency and the number of times latency was registered.

At this point, basic functionality exists, along with one
test.