add some sync-related tools, e.g safe chan iter and send
The planned monitor implementation will do a man-in-the-middle
injection for KNNEnqueueResults.Pipe (chan) and will do some
risky chan reads and writes. To mitigate leaks through flawed
implementation, some safety funcs are defined here.

The safety is in the form of chan reads and writes that are
protected with context.Context (so deadlines can be used).

syncutils.go (new):
        - safeChanIter(...)
        - safeChanSend(...)
        - ctxDone(...)