started ./service/ops pkg (rpc on top of requestman.Handle)
This pkg will act as an rpc layer on top of requestman.Handle.
This commit adds a Client and a Server type which do some
basic rpc setup and connectivity routines (i.e the very basics
of rpc in Go). Additionally, the following requestman.Handle
methods are accounted for now:
- AddData
- KNN
This is mainly in client.go and server.go.

They both are tested (there is an additional test for Ping).
This is mainly in client_test.go

Besides this, some testing utils are added, for doing things
such as:
- Finding a free port
- Random (valid) requestman.KNNArgs
- Convenient and consistent setup of requestman.Handle
- Convenient and consistent setup of aforementioned Server.
- Simplicity of tests with boiler-plate reduction
Mainly in testutils.go