first api commit; Server, utils + ping test
First commit for the pkg that will be a JSON/POSt server on top
of /service/ops. As an overview, this commit adds the server,
config, and an endpoint for ping. Additionally, some testing-
and io tools are added.

api.go:
        T handle: Storing endpoints and state.
        M handle.registerRoutes: Registering routes.
        T StartServerArgs: (cfg) for F StartServer below.
        F StartServer.

api_test.go:
        F TestPing: Test Ping endpoint (check routes.go).

ioconv.go:
        F withNetIO: Easy generic json encoding/decoding

routes.go:
        M handle.Ping: Endpoint for ping.

testutils.go:
        F freePort: Getting a free os port.
        F freePortNoFail: Similar to the above.
        F post: Easy generic http.Post.